samples: sensor: lm75: sample for lm75 added
This adds a sample for using a LM75 temperature sensor.
It is based on the adt7420 sample, modified and extended
accordingly to work with basic LM75 sensor. For testing,
the Atmel AT30TS75A was used, which allows advanced
features such as user-configurable resolution.
The lm75 driver was thus extended by an attribute-set function,
which allows to set the configuration register of the sensor.

Signed-off-by: Loris Schmit <loris.schmit@inovex.de>